TIP-860: Move AttributeGroup related api tests
