+++ tests/back/Pim/Structure/EndToEnd/AttributeGroup/ExternalApi/AttributeGroupAuthorizationEndToEnd.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Security;
+namespace AkeneoTest\Pim\Structure\EndToEnd\AttributeGroup\ExternalApi;
 
 use Akeneo\Test\Integration\Configuration;
 use Akeneo\Tool\Bundle\ApiBundle\Stream\StreamResourceResponse;
 use Akeneo\Tool\Bundle\ApiBundle\tests\integration\ApiTestCase;
 use Symfony\Component\HttpFoundation\Response;
 
-class AttributeGroupAuthorizationIntegration extends ApiTestCase
+class AttributeGroupAuthorizationEndToEnd extends ApiTestCase
 {
     public function testOverallAccessDenied()
     {
