+++ tests/back/Pim/Structure/EndToEnd/AttributeGroup/ExternalApi/PartialUpdateAttributeGroupEndToEnd.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller\AttributeGroup;
+namespace AkeneoTest\Pim\Structure\EndToEnd\AttributeGroup\ExternalApi;
 
 use Akeneo\Test\Integration\Configuration;
 use Akeneo\Tool\Bundle\ApiBundle\tests\integration\ApiTestCase;
@@ -9,7 +9,7 @@ use Symfony\Component\HttpFoundation\Response;
 /**
  * @group ce
  */
-class PartialUpdateAttributeGroupIntegration extends ApiTestCase
+class PartialUpdateAttributeGroupEndToEnd extends ApiTestCase
 {
     public function testHttpHeadersInResponseWhenAnAttributeGroupIsUpdated()
     {
